requirements.txt
numpy>=1.24
opencv-python>=4.8
Pillow>=10.0
click>=8.1
scipy>=1.11
onnxruntime>=1.18
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
joblib>=1.3.0
albumentations>=1.3.0
ultralytics>=8.1.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
requests>=2.32.0
